+++ CHANGELOG.md
@@ -12,6 +12,11 @@ This project attempts to adhere to [Semantic Versioning](http://semver.org).
 
 - reduce binary size of Levenshtein distance
 
+- improve Damerau-Levenshtein implementation
+  - reduce memory usage from `O(N*M)` to `O(N+M)`
+  - reduce runtime in our own benchmark by more than `70%`
+  - reduce binary size by more than `25%`
+
 ### Fixed
 
 - Fix transposition counting in Jaro and Jaro-Winkler.